# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoPie API documentation",
        "title": "GoPie API",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/source/s3/upload": {
            "post": {
                "description": "Upload a file from S3 and create a new dataset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "s3"
                ],
                "summary": "Upload file from S3",
                "parameters": [
                    {
                        "description": "Upload request parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/s3.uploadRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Dataset"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or S3 file access error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Dataset": {
            "description": "Dataset model",
            "type": "object",
            "properties": {
                "alias": {
                    "description": "Alias of the dataset",
                    "type": "string",
                    "example": "sales_data_alias"
                },
                "columns": {
                    "description": "Column definitions of the dataset",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                },
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string",
                    "example": "2024-02-05T12:00:00Z"
                },
                "created_by": {
                    "description": "CreatedBy represents the user who created the dataset",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "description": {
                    "description": "Description of the dataset",
                    "type": "string",
                    "example": "Sales data for Q1 2024"
                },
                "file_path": {
                    "description": "File path of the dataset",
                    "type": "string",
                    "example": "/data/sales_data.csv"
                },
                "format": {
                    "description": "Format of the dataset (e.g., csv, parquet)",
                    "type": "string",
                    "example": "csv"
                },
                "id": {
                    "description": "Unique identifier of the dataset",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "description": "Name of the dataset",
                    "type": "string",
                    "example": "gp_Dh790Asdf17kd"
                },
                "row_count": {
                    "description": "Number of rows in the dataset",
                    "type": "integer",
                    "example": 1000
                },
                "size": {
                    "description": "Size of the dataset in bytes",
                    "type": "integer",
                    "example": 1048576
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2024-02-05T12:00:00Z"
                },
                "updated_by": {
                    "description": "UpdatedBy represents the user who last updated the dataset",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "responses.ErrorResponse": {
            "description": "Error response structure",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "description": "Error details",
                    "type": "string",
                    "example": "Validation failed"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
        "responses.SuccessResponse": {
            "description": "Success response structure with data",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data"
                }
            }
        },
...
        "s3.uploadRequestBody": {
            "description": "Request body for uploading a file from S3",
            "type": "object",
            "required": [
                "alias",
                "alter_column_names",
                "column_descriptions",
                "created_by",
                "file_path",
                "project_id"
            ],
            "properties": {
                "alias": {
                    "description": "Alias of the dataset",
                    "type": "string",
                    "minLength": 3,
                    "example": "sales_data"
                },
                "alter_column_names": {
                    "description": "Column names to be altered",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "column_descriptions": {
                    "description": "Column descriptions",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "created_by": {
                    "description": "User ID of the creator",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "description": {
                    "description": "Description of the dataset",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 10,
                    "example": "Sales data for Q1 2024"
                },
                "file_path": {
                    "description": "S3 path of the file to upload",
                    "type": "string",
                    "minLength": 1,
                    "example": "my-bucket/data/sales.csv"
                },
                "project_id": {
                    "description": "ID of the project to add the dataset to",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        }
    }
}